Separate out dynamic value utils
